+++ ledger/ledgerpayout_genmain.go
@@ -1,0 +1,5 @@
+// This file is auto generate
+package ledger
+
+func initMain() {
+}
